requirements.txt
Django==5.1.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
openai==1.3.7
python-dotenv==1.0.0
requests==2.32.3
gunicorn==21.2.0
whitenoise==6.6.0 